Resolve "[RPC] public/get_tradingview_chart_data"

+++ internal/deribit/model/deribit.go
@@ -534,12 +534,12 @@ type GetTradingviewChartDataRequest struct {
 }
 
 type GetTradingviewChartDataResponse struct {
-	Close  []int64 `json:"close"`
-	Cost   []int64 `json:"cost"`
-	High   []int64 `json:"high"`
-	Low    []int64 `json:"low"`
-	Open   []int64 `json:"open"`
-	Tics   []int64 `json:"tics"`
-	Volume []int64 `json:"volume"`
-	Status string  `json:"status"`
+	Close  []float64 `json:"close"`
+	Cost   []float64 `json:"cost"`
+	High   []float64 `json:"high"`
+	Low    []float64 `json:"low"`
+	Open   []float64 `json:"open"`
+	Tics   []int64   `json:"tics"`
+	Volume []float64 `json:"volume"`
+	Status string    `json:"status"`
 }
